Refactor file downloading.
The nxm url is now read into a struct that gets passed around, instead
of every function taking a ton of parameters. Since the url parsing
could be moved to the NxmUrl module, the request module didn't get
significantly more crowded by getting rid of the nxmhandler module.

The db is renamed to cache, since that's a more descriptive name.