Implement md5 lookup of mod archive.
It looks like a lot of refactoring was involved here, but the gist of it
is that cache-related logic is moved from api/request.rs to lookup.rs,
since the request module should be oblivious to the cache. Additionally,
the different structs to manage API responses were starting to pile up,
so they were moved to their own directory.