Added logging and updated for async/await
This commit changes a lot and that's not ideal. It combines a lot of
small refactoring done during what's essentially a hiatus from
development. The project is still undergoing major changes and I'm
figuring out Rust as I go. I'm certain there are still bugs, and thus
far the program is just a nxm url handler that can do useless API
requests.

Updating the dependencies meant having to adapt the new async/await
syntax, which causes edits throughout the program.

This commit bundles a new logging implementation with the "log"
crate which was helpful when debugging the other changes. Setting the
log level to "debug" or "trace" also shows messages from libraries,
which can be convenient.

Command line argument parsing has been moved to the cmd module to
prevent crowding the main module. It might move back if args can be
declared concisely enough with some of clap's other methods for
declaring args.