Implement custom error types
And a bunch of unrelated small changes...

The lookup module's code was a bunch of error handling spaghetti that
branched a bunch just to be able to print the appropriate error
messages. This was before I learned to use Box<dyn Error>.

Using custom error types instead of boxed errors allows explicitly
defining the expected failure cases for an operation and propagating
errors like "ApiKeyMissing" from the backend to the frontend.
This approach might turn out to be a mistake, and writing impls for the
errors certainly adds a lot of boilerplate.

This commit includes the first unit tests. The unit tests necessarily
include data to test serialization with. I've used actual data from a
popular mod. I'm not a fan of adding it to the repo, but the data was
already public.