Rename error types and clean up code
DownloadError is renamed to RequestError and NxmDownloadError is now
DownloadError.

Downloading broke because the timestamp in the nxm download *seems* to
have changed format from milliseconds to seconds, and this fixes that.
Timestamps are now in seconds since epoch.

There's also some small code cleanups and bug fixes caused by everything
changing all the time.