Move requests to Client and reshuffle modules
A lot of refactoring here. Requests need to use a shared client in order
to reuse connections. This currently makes the Queriable trait slightly
unwieldy and it might face a rewrite later.

The lookup module will probably be deprecated soonish because its
responsibilities are overtaken by Client and Cache.

The cacheable trait was moved under db.

A lot of possibility for breakage here, but it compiles.