Refactor UI due to multithreading reqs
After a lot of hair-pulling with Arc, RwLock and HashMap I decided to
split StatefulCollection into just the State tracking. The state only
needs to know about the length of the collection, and that can be passed
as a command line parameter. It's unwieldy to use, but solves a bunch of
other problems for now.

The UI now takes Cache and Client as parameters, which makes things
a lot cleaner.

The UI can now poll the cache for file changes so we don't need to
redraw the table all the time. The same will be done for downloads later
on.