Major UI rewrite. Details below.
The UI code became rather bloated to work with so it was time to split
it up. The primary objective was to simplify writing the code for
keybinds and UI state changing, but I succeeded only partially.

Component creation code moved to component/widget. The previous state
traits are unified in the Select trait. There's a new Highlight trait
that applies different styles to the focused block.

I'm quit unhappy with the way things ended up, but they're better than
before. I ended up using macros to work around limitations of the trait
system. The solution is "fine" apart from ideological reasons. The
mindless boilerplate responsible for creating widgets might meet a
similar fate later on.