Clean up some modules a bit, details in commit msg
The update module is moved from cache to api since it seems better
suited there.
The UpdateError type was removed since it was unused. This left the
error directory with only one module, so the hierarchy was flattened.

The download modules were moved in order to prevent a redundant
directory structure like download/downloads.rs.