Even more Cache refactoring
The Cache initialization logic was really hard to follow. Splitting the
initialization into separate modules allowed me to better understand
what it's supposed to do, which led to more refactoring elsewhere.

The cause of the original complexity is that the UI requires a bunch of
information about the files. LocalFile doesn't contain all necessary
data, and I don't think constantly performing lookups to FileListCache
is a very good solution. I created a separate collection for those
FileDetails we're actually interested in, then got tangled up in the
logic.