Avoid needlessly recalculating UI bounds
The UI is finally becoming pretty snappy. The main thread still freezes
when mass checking for updates, and there are some probable race
conditions, but there's definitely been improvement.

I left the UI in a broken state after the last refactor (I learned that
selectively redrawing the screen is seemingly not possible).
Highlighting is still broken,