Merge LocalFileCache into FileIndex
Turns out that LocalFile and FileDetails are needed basically in
parallel. The FileDetailsCache and LocalFileCache are now merged into
one, which makes drawing the file list a lot easier. I also got rid of
all the places where the LocalFile is cloned, and now the UI is able to
display the update status (apart from the update not triggering a
redraw).