Rework config after SSO was added
The config logic got overhauled because providing the API key before
starting the program is no longer necessary.
Turns out that setting the game in the config file is no longer
necessary either - it just creates a subdirectory inside the download
directory. However, this can be convenient to users and it makes unit
tests simpler, so I renamed it to "profile", which hopefully describes
well enough what the feature is about.

The rest is small changes to places that were affected, apart from
main.rs, where the initialization could finally be made simple.