Fix crash when resizing terminal
I was handling sigwinch and recalculating Rects at the beginning of the
UI loop, while checking terminal size by calling termion directly. This
caused a race condition because the Rect sizes were often outdated by
the time Terminal::draw() is called. As a side effect of the fix the
code became a bit cleaner.