Internal data structure revamp
Start saving and using Md5Results since it's a convenient way to get a
ModInfo, which contains things like the mod name which we wouldn't
otherwise know.

FileLists were bloated with a lot of extra file versions and update
history. Pruning these after downloading them makes everything else faster.
In order to do so we need to know the lowest file id and earlier timestamp,
which the previous BinaryHeap implementation was not suitable for.
Instead revert to using a vec which takes O(n) to verify is sorted.
This is probably less work than serde does constructing the binary heap
anyway.

While at it, wrap the FileDetails in FileData inside Option since they're
not guaranteed to be present (for network related reasons, perhaps). The
update checking still contains some unwraps, but other places now handle
missing FileDetails gracefully.

And finally remove the RwLock on LocalFile because it's used everywhere
and the horrible API was even bleeding into UI code. Instead create a
wrapper for update_status that exposes an enum while using atomics
internally. It still needs to (de)serialize as an enum for backward
compatibility reasons, which caused an interesting side quest with no
quest reward.

This could have been several commits, but splitting it all up is not time
well spent.