Unreasonably destructive backend rewrite
For sanity reasons I won't try to split this up or fix every minor broken
thing. I sure wish I had more unit tests for all this.

Until now dmodman hasn't had any metadata for extracted files, nor has it
shown them in the UI. Things that should really be done for *installed*
mods (like update checking) were being done based on the metadata of
downloaded mods.

The file index (renamed to metadata index) had the dual purpose of
deserializing the metadata of archives and providing various ways to look
up mod metadata. This is now finally split up into modules that each
serve a fairly clear purpose, but it touches a *lot* of code. The metadata
index uses a new type ModFileMetadata which contains everything we know
about a mod's file (and its related mod), but it gets concurrently read and
written to as new data is fetched, so there's locks absolutely everywhere.
This avoids a lot of hashmap lookups, but it's unwieldy to use.

The ModFileMetadata has a one-to-many relationship with both archives and
installed mods (an archive can have multiple copies under different names
or be installed multiple times with different settings). The MFM is kept
updated by the MetadataIndex, and the logic there could use some testing
and improving. Still, it's about time that all this gets committed before
I break something again.

The test files are moved to a differently named subdirectory because they
used the by-now-almost-ancient logic where the modded games and profiles
were inseparable. It needed fixing anyways, but it became confusing as I
started thinking about writing unit tests for all this.