Uncouple UI selected index from backend index
There was a longstanding race condition where the UI was directly using
indices in the backend, but those could be modified outside of UI code (
for example a download finishing and being added to the archive list).
Now each UI widget remembers the last shown UI state, and the frontend code
can use that to get the proper identifier for things in the cache.

Fix some minor things like renaming FileTable to InstalledModsTable
since there are now multiple types of files in the UI.

Add missing hotkeys for common functions that only had keybindings in one
of the UI widgets.