API, metadata and update things
As usual, too much stuff in one commit. I hope to stabilize things once the
things I've written for this release are working.

All this metadata and state handling is turning into quite a bit of work.
I still think the approach is sound enough, but I end up doing a lot of
iteration on the code until it starts making sense.

Notable changes:
- Stop using two file details types (md5results and file details) because
it causes ugly code while only making the program a bit more fault reliant
when meta data that shouldn't be missing anyways goes missing
- Refactor code that queries the API into its own module so that cache
updates are properly handled
- Add keybind to query missing metadata for a mod.
- Wrap the config in an Arc. I've avoided doing this because of how much
code it touches and the risk of having to revert it some day, but it was
copied a lot. Wrapping each field in an arc would reduce side effects but
would multiply the reference counters only for code ergonomics.