Change UI element placement
Having the log below other elements takes up screen space and forces me to
come up with a way to intuitively navigate to the element below and back.
Changing the UI design a bit seems like a saner approach. Seeing archives
and downloads first should be more useful to people who don't want to use
downloader features.

Also, turns out the API I'm exposing in the UI is still a mess and
reordering things required a lot of changes in random places.